+++ emr-logservice/src/main/scala/org/apache/spark/streaming/aliyun/logservice/writer/CachedProducer.scala
@@ -1,4 +1,20 @@
-package org.apache.spark.streaming.aliyun.logservice
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.spark.streaming.aliyun.logservice.writer
 
 import com.aliyun.openservices.aliyun.log.producer._
 import com.aliyun.openservices.log.common.LogItem
@@ -41,16 +57,18 @@ object CachedProducer {
       val logProducer = new LogProducer(config)
       logProducer.putProjectConfig(new ProjectConfig(logProject, endpoint, accessKeyId, accessKeySecret))
       producer = new ProducerAdapter(logProject, logstore, logProducer)
+      sys.addShutdownHook(producer.close())
     }
     producer
   }
 
   /**
-   * Flush and close the [[Producer]] in the cache associated with this config
-   *
-   * @param producerConfig producer configuration associated to a [[Producer]]
+   * Flush and close the cached [[Producer]]
    */
-  def close(producerConfig: Map[String, Object]): Unit = {
-    producer.close()
+  def close(): Unit = {
+    if (producer != null) {
+      producer.close()
+      producer = null
+    }
   }
 }
